Consolidate meta programming utilities
